refactor x86_64
