+++ crabml-core/src/backends/cpu/archutil/mod.rs
@@ -1,3 +1,6 @@
 #[allow(dead_code)]
 #[cfg(target_arch = "aarch64")]
 pub mod aarch64;
+
+#[cfg(target_arch = "x86_64")]
+pub mod x86_64;
