+++ crabml-core/src/backends/cpu/buf/buf_q8_0.rs
@@ -43,7 +43,7 @@ impl<'a> QuantBufQ8_0<'a> {
     }
 
     pub fn quantize(data: &[f32]) -> Self {
-        let bs = quantize_f32_q8_0_stdsimd(data);
+        let bs = quantize_f32_q8_0(data);
         Self { blocks: bs.into() }
     }
 
@@ -78,7 +78,7 @@ impl<'a> QuantBufQ8_0<'a> {
     }
 }
 
-pub fn quantize_f32_q8_0_stdsimd(data: &[f32]) -> Vec<BlockQ8_0> {
+pub fn quantize_f32_q8_0(data: &[f32]) -> Vec<BlockQ8_0> {
     use std::simd::f32x4;
 
     let mut bs = Vec::with_capacity(data.len() / 32);
@@ -134,7 +134,7 @@ fn vec_dot_q8_0_q8_0(abs: &[BlockQ8_0], bbs: &[BlockQ8_0]) -> f32 {
 
     #[cfg(all(target_arch = "x86_64", target_feature = "avx2"))]
     {
-        return impl_x86_64_avx2::vec_dot_q8_0_q8_0_avx2(abs, bbs);
+        return vec_dot_q8_0_q8_0_avx2(abs, bbs);
     }
 
     #[cfg(not(any(
@@ -147,12 +147,6 @@ fn vec_dot_q8_0_q8_0(abs: &[BlockQ8_0], bbs: &[BlockQ8_0]) -> f32 {
 #[cfg(all(target_arch = "aarch64", target_feature = "neon"))]
 fn vec_dot_q8_0_q8_0_neon(abs: &[BlockQ8_0], bbs: &[BlockQ8_0]) -> f32 {
     use std::arch::aarch64;
-
-    assert!(
-        bbs.len() % 2 == 0,
-        "bbs.len() must be a multiple of 64, got: {}",
-        bbs.len()
-    );
 
     let blocks_rounded = bbs.len() - bbs.len() % 2;
     let mut result = 0.0;
@@ -223,95 +217,48 @@ fn vec_dot_q8_0_q8_0_neon(abs: &[BlockQ8_0], bbs: &[BlockQ8_0]) -> f32 {
 }
 
 #[cfg(all(target_arch = "x86_64", target_feature = "avx2"))]
-mod impl_x86_64_avx2 {
+pub fn vec_dot_q8_0_q8_0(abs: &[BlockQ8_0], bbs: &[BlockQ8_0]) -> f32 {
     use std::arch::x86_64::*;
 
-    use half::f16;
+    use crate::backends::cpu::archutil::x86_64::*;
+    debug_assert_eq!(abs.len(), bbs.len());
 
-    use super::BlockQ8_0;
+    unsafe {
+        let mut acc0 = _mm256_setzero_ps();
+        let mut acc1 = _mm256_setzero_ps();
 
-    /// Inspired a lot by ggml https://github.com/ggerganov/ggml/blob/master/src/ggml-quants.c
-    pub fn vec_dot_q8_0_q8_0(abs: &[BlockQ8_0], bbs: &[BlockQ8_0]) -> f32 {
-        debug_assert_eq!(abs.len(), bbs.len());
+        for [(abs0, bbs0), (abs1, bbs1)] in abs.iter().zip(bbs).array_chunks::<2>() {
+            let d0 = _mm256_set1_ps(abs0.d.to_f32() * bbs0.d.to_f32());
+            let d1 = _mm256_set1_ps(abs1.d.to_f32() * bbs1.d.to_f32());
 
-        unsafe {
-            let mut acc0 = _mm256_setzero_ps();
-            let mut acc1 = _mm256_setzero_ps();
+            let qa0 = _mm256_loadu_si256(abs0.qs.as_ptr() as *const __m256i);
+            let qb0 = _mm256_loadu_si256(bbs0.qs.as_ptr() as *const __m256i);
 
-            for [(abs0, bbs0), (abs1, bbs1)] in abs.iter().zip(bbs).array_chunks::<2>() {
-                let d0 = _mm256_set1_ps(abs0.d.to_f32() * bbs0.d.to_f32());
-                let d1 = _mm256_set1_ps(abs1.d.to_f32() * bbs1.d.to_f32());
+            let qa1 = _mm256_loadu_si256(abs1.qs.as_ptr() as *const __m256i);
+            let qb1 = _mm256_loadu_si256(bbs1.qs.as_ptr() as *const __m256i);
 
-                let qa0 = _mm256_loadu_si256(abs0.qs.as_ptr() as *const __m256i);
-                let qb0 = _mm256_loadu_si256(bbs0.qs.as_ptr() as *const __m256i);
+            let q0 = mul_sum_i8_pairs_float(qa0, qb0);
+            let q1 = mul_sum_i8_pairs_float(qa1, qb1);
 
-                let qa1 = _mm256_loadu_si256(abs1.qs.as_ptr() as *const __m256i);
-                let qb1 = _mm256_loadu_si256(bbs1.qs.as_ptr() as *const __m256i);
-
-                let q0 = mul_sum_i8_pairs_float(qa0, qb0);
-                let q1 = mul_sum_i8_pairs_float(qa1, qb1);
-
-                acc0 = _mm256_fmadd_ps(d0, q0, acc0);
-                acc1 = _mm256_fmadd_ps(d1, q1, acc1);
-            }
-
-            if abs.len() % 2 == 1 {
-                let a = abs.last().unwrap_unchecked();
-                let b = abs.last().unwrap_unchecked();
-
-                let d = _mm256_set1_ps(a.d.to_f32() * b.d.to_f32());
-
-                let qa = _mm256_loadu_si256(a.qs.as_ptr() as *const __m256i);
-                let qb = _mm256_loadu_si256(b.qs.as_ptr() as *const __m256i);
-
-                let q = mul_sum_i8_pairs_float(qa, qb);
-
-                acc0 = _mm256_fmadd_ps(d, q, acc0);
-            }
-
-            hsum_float_8(_mm256_add_ps(acc0, acc1))
+            acc0 = _mm256_fmadd_ps(d0, q0, acc0);
+            acc1 = _mm256_fmadd_ps(d1, q1, acc1);
         }
-    }
 
-    /// TODO: Adding AVX-VNNI support so that we can use `_mm256_dpbssd_epi32`
-    #[inline]
-    unsafe fn mul_sum_i8_pairs_float(x: __m256i, y: __m256i) -> __m256 {
-        // Get absolute values of x vectors
-        let ax = _mm256_sign_epi8(x, x);
-        // Sign the values of the y vectors
-        let sy = _mm256_sign_epi8(y, x);
-        mul_sum_us8_pairs_float(ax, sy)
-    }
+        if abs.len() % 2 == 1 {
+            let a = abs.last().unwrap_unchecked();
+            let b = abs.last().unwrap_unchecked();
 
-    #[inline]
-    unsafe fn mul_sum_us8_pairs_float(ax: __m256i, sy: __m256i) -> __m256 {
-        let axl = _mm256_castsi256_si128(ax);
-        let axh = _mm256_extractf128_si256(ax, 1);
-        let syl = _mm256_castsi256_si128(sy);
-        let syh = _mm256_extractf128_si256(sy, 1);
-        // Perform multiplication and create 16-bit values
-        let dotl = _mm_maddubs_epi16(axl, syl);
-        let doth = _mm_maddubs_epi16(axh, syh);
-        sum_i16_pairs_float(doth, dotl)
-    }
+            let d = _mm256_set1_ps(a.d.to_f32() * b.d.to_f32());
 
-    #[inline]
-    unsafe fn sum_i16_pairs_float(xh: __m128i, xl: __m128i) -> __m256 {
-        let ones = _mm_set1_epi16(1);
-        let summed_pairsl = _mm_madd_epi16(ones, xl);
-        let summed_pairsh = _mm_madd_epi16(ones, xh);
-        let summed_pairs = _mm256_set_m128i(summed_pairsh, summed_pairsl);
-        _mm256_cvtepi32_ps(summed_pairs)
-    }
+            let qa = _mm256_loadu_si256(a.qs.as_ptr() as *const __m256i);
+            let qb = _mm256_loadu_si256(b.qs.as_ptr() as *const __m256i);
 
-    /// horizontally add 8 floats
-    #[inline]
-    unsafe fn hsum_float_8(x: __m256) -> f32 {
-        let res = _mm256_extractf128_ps(x, 1);
-        let res = _mm_add_ps(res, _mm256_castps256_ps128(x));
-        let res = _mm_add_ps(res, _mm_movehl_ps(res, res));
-        let res = _mm_add_ss(res, _mm_movehdup_ps(res));
-        _mm_cvtss_f32(res)
+            let q = mul_sum_i8_pairs_float(qa, qb);
+
+            acc0 = _mm256_fmadd_ps(d, q, acc0);
+        }
+
+        hsum_float_8(_mm256_add_ps(acc0, acc1))
     }
 }
